move calculateTheClosestNextRowKeyForPrefix to own util
